+++ packages/clima_ui/lib/screens/settings_screen.dart
@@ -2,9 +2,11 @@ import 'package:clima_data/models/dark_theme_model.dart';
 import 'package:clima_data/models/theme_model.dart';
 import 'package:clima_ui/screens/about_screen.dart';
 import 'package:clima_ui/state_notifiers/theme_state_notifier.dart';
-import 'package:clima_ui/widgets/dark_theme_dialog.dart';
-import 'package:clima_ui/widgets/reusable_widgets.dart';
-import 'package:clima_ui/widgets/theme_dialog.dart';
+import 'package:clima_ui/widgets/dialogs/dark_theme_dialog.dart';
+import 'package:clima_ui/widgets/dialogs/theme_dialog.dart';
+import 'package:clima_ui/widgets/settings/settings_divider.dart';
+import 'package:clima_ui/widgets/settings/settings_header.dart';
+import 'package:clima_ui/widgets/settings/settings_tile.dart';
 // import 'package:clima_ui/widgets/unit_dialog.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
