+++ packages/clima_ui/lib/widgets/current_conditions.dart
@@ -1,12 +1,10 @@
 import 'package:auto_size_text/auto_size_text.dart';
 import 'package:clima_ui/state_notifiers/weather_state_notifier.dart';
 import 'package:clima_ui/utilities/constants.dart';
-import 'package:clima_ui/utilities/weather_icons.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:sizer/sizer.dart';
-import 'package:weather_icons/weather_icons.dart';
 
 import 'value_tile.dart';
 
@@ -64,32 +62,32 @@ class CurrentConditions extends HookWidget {
             ),
           ),
         ),
-        Expanded(
-          flex: 4,
-          child: BoxedIcon(
-            getIconData(weather.iconCode),
-            color: Theme.of(context).textTheme.subtitle1!.color,
-            size: () {
-              switch (weather.iconCode) {
-                case '03d':
-                case '04d':
-                case '03n':
-                case '04n':
-                case '01n':
-                case '01d':
-                  return MediaQuery.of(context).size.shortestSide <
-                          kTabletBreakpoint
-                      ? 68.75.sp
-                      : 51.75.sp;
-                default:
-                  return MediaQuery.of(context).size.shortestSide <
-                          kTabletBreakpoint
-                      ? 60.sp
-                      : 45.sp;
-              }
-            }(),
-          ),
-        ),
+        // Expanded(
+        //   flex: 4,
+        //   child: BoxedIcon(
+        //     getIconData(weather.iconCode),
+        //     color: Theme.of(context).textTheme.subtitle1!.color,
+        //     size: () {
+        //       switch (weather.iconCode) {
+        //         case '03d':
+        //         case '04d':
+        //         case '03n':
+        //         case '04n':
+        //         case '01n':
+        //         case '01d':
+        //           return MediaQuery.of(context).size.shortestSide <
+        //                   kTabletBreakpoint
+        //               ? 68.75.sp
+        //               : 51.75.sp;
+        //         default:
+        //           return MediaQuery.of(context).size.shortestSide <
+        //                   kTabletBreakpoint
+        //               ? 60.sp
+        //               : 45.sp;
+        //       }
+        //     }(),
+        //   ),
+        // ),
         Expanded(
           flex: 3,
           child: AutoSizeText(
