+++ packages/clima_ui/lib/widgets/settings/settings_divider.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class SettingsDivider extends StatelessWidget {
+  const SettingsDivider({
+    Key? key,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Divider(
+      height: 1,
+      color: Theme.of(context).textTheme.subtitle1!.color!.withAlpha(65),
+    );
+  }
+}
