+++ packages/clima_ui/lib/widgets/settings/settings_header.dart
@@ -1,0 +1,24 @@
+import 'package:flutter/material.dart';
+
+class SettingsHeader extends StatelessWidget {
+  const SettingsHeader({required this.title});
+  final String title;
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.only(top: 24),
+      child: Container(
+        height: 30.0,
+        padding: const EdgeInsets.symmetric(horizontal: 80),
+        alignment: Alignment.centerLeft,
+        child: Text(
+          title,
+          style: Theme.of(context).textTheme.bodyText1!.copyWith(
+                color: Theme.of(context).accentColor,
+              ),
+        ),
+      ),
+    );
+  }
+}
