+++ packages/clima_ui/pubspec.yaml
@@ -25,12 +25,12 @@ dependencies:
   flutter_hooks: 0.17.0
   flutter_riverpod: 0.14.0+3
   flutter_spinkit: 5.0.0
-  weather_icons: 3.0.0
   hooks_riverpod: 0.14.0+4
   material_floating_search_bar:
     git:
       url: https://github.com/CentaurusApps/material_floating_search_bar.git
       ref: allow-passing-system-overlay-style
+  font_awesome_flutter: 9.1.0
   package_info: 2.0.2
   riverpod: 0.14.0+3
   shared_preferences: 2.0.6
